remove int usage for rlp

+++ nimbus/sync/handlers/eth.nim
@@ -448,7 +448,7 @@ method handleAnnouncedTxsHashes*(
       ctx: EthWireRef;
       peer: Peer;
       txTypes: Blob;
-      txSizes: openArray[int];
+      txSizes: openArray[uint64];
       txHashes: openArray[Hash256];
         ): Result[void, string] =
   when extraTraceMessages:
